Add repositories and fix bugs

+++ src/main/java/pl/ssanko/petclinic/data/entity/MedicalProcedure.java
@@ -3,6 +3,8 @@ package pl.ssanko.petclinic.data.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @Table(name = "medical_procedures")
@@ -12,6 +14,6 @@ public class MedicalProcedure {
     private Long id;
     private String name;
     private String description;
-    private float cost;
+    private BigDecimal cost;
 
 }
